Refactor game loop, extract main logic to game package, add spawn points
